# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Indonesia wildfire hotspot analytics (detections, GeoJSON, summary, drill-down locations and periods).",
        "title": "Hotspot API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/hotspots/": {
            "get": {
                "description": "Retrieve a paginated list of raw hotspot detections.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotspots"
                ],
                "summary": "List hotspots",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max records to return (1-100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response-domain_GetHotspotsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hotspots/filter-options": {
            "get": {
                "description": "Distinct confidence classes, satellite names, and products (NRT/SP) for filter dropdowns.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotspots"
                ],
                "summary": "Filter options",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response-domain_GetFilterOptionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hotspots/geojson": {
            "get": {
                "description": "Retrieve hotspot detections as a GeoJSON FeatureCollection with cursor pagination. Supports date-range, time-period, location and product (NRT/SP) filters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotspots"
                ],
                "summary": "Hotspots as GeoJSON",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by year",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by semester",
                        "name": "semester",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by quarter",
                        "name": "quarter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by month",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by week",
                        "name": "week",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by province code",
                        "name": "province_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by city code",
                        "name": "city_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by district code",
                        "name": "district_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by subdistrict code",
                        "name": "subdistrict_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by satellite name",
                        "name": "satellite",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by confidence class",
                        "name": "confidence",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "NRT",
                            "SP"
                        ],
                        "type": "string",
                        "description": "Filter by product",
                        "name": "product",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor from previous page",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max features to return (1-50000, default 500)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response-domain_GeoJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hotspots/locations": {
            "get": {
                "description": "Drill-down geographic hierarchy with hotspot counts. Returns islands with grouped provinces at the top level, or cities/districts/subdistricts as the location code is drilled down.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotspots"
                ],
                "summary": "Location hierarchy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Drill into this province",
                        "name": "province_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Drill into this city",
                        "name": "city_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Drill into this district",
                        "name": "district_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by satellite name",
                        "name": "satellite",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by confidence class",
                        "name": "confidence",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "NRT",
                            "SP"
                        ],
                        "type": "string",
                        "description": "Filter by product",
                        "name": "product",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by year",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by semester",
                        "name": "semester",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by quarter",
                        "name": "quarter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by month",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by week",
                        "name": "week",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response-domain_GetLocationsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hotspots/periods": {
            "get": {
                "description": "Drill-down time period options (years then semesters, quarters, months, weeks) based on the params supplied.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotspots"
                ],
                "summary": "Time period options",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Selected year",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Selected semester",
                        "name": "semester",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Selected quarter",
                        "name": "quarter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Selected month",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response-domain_GetPeriodsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hotspots/summary": {
            "get": {
                "description": "Aggregated dashboard summary: top provinces/cities, satellite and confidence distribution, overall stats, monthly stats and today's stats.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotspots"
                ],
                "summary": "Dashboard summary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Top provinces limit (1-100, default 10)",
                        "name": "province_limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Top cities limit (1-100, default 10)",
                        "name": "city_limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response-domain_GetSummaryResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "domain.FilterOption": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "domain.GetFilterOptionsResponse": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.FilterOption"
                    }
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.FilterOption"
                    }
                },
                "satellites": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.FilterOption"
                    }
                }
            }
        },
        "domain.GetHotspotsResponse": {
            "type": "object",
            "properties": {
                "hotspots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.HotspotDetail"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/domain.Pagination"
                }
            }
        },
...
        "domain.Response-domain_GeoJSON": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.GeoJSON"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "domain.Response-domain_GetFilterOptionsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.GetFilterOptionsResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "domain.Response-domain_GetHotspotsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.GetHotspotsResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "domain.Response-domain_GetLocationsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.GetLocationsResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "domain.Response-domain_GetPeriodsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.GetPeriodsResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "domain.Response-domain_GetSummaryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.GetSummaryResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}